Update README files for better documentation
The two README.md files are updated to better document this plugin.

The main README.md file is updated to point to the README.md file
inside the functions folder to read about the documentation.

Inline code blocks where use to proper differentiate the commands from
the rest of the text